+++ bcdi/preprocessing/bcdi_utils.py
@@ -1183,65 +1183,3 @@ def reload_bcdi_data(
         )
 
     return data, mask, frames_logical, monitor
-
-
-def zero_pad(array, padding_width=np.zeros(6), mask_flag=False, debugging=False):
-    """
-    Pad obj with zeros.
-
-    :param array: 3D array to be padded
-    :param padding_width: number of zero pixels to padd on each side
-    :param mask_flag: set to True to pad with 1, False to pad with 0
-    :type mask_flag: bool
-    :param debugging: set to True to see plots
-    :type debugging: bool
-    :return: obj padded with zeros
-    """
-    valid.valid_ndarray(arrays=array, ndim=3)
-    nbz, nby, nbx = array.shape
-    if all(x == 0 for x in np.zeros(6)):
-        return array
-
-    if debugging:
-        gu.multislices_plot(
-            array=array,
-            sum_frames=False,
-            plot_colorbar=True,
-            vmin=0,
-            vmax=1,
-            title="Array before padding",
-        )
-
-    if mask_flag:
-        newobj = np.ones(
-            (
-                nbz + padding_width[0] + padding_width[1],
-                nby + padding_width[2] + padding_width[3],
-                nbx + padding_width[4] + padding_width[5],
-            )
-        )
-    else:
-        newobj = np.zeros(
-            (
-                nbz + padding_width[0] + padding_width[1],
-                nby + padding_width[2] + padding_width[3],
-                nbx + padding_width[4] + padding_width[5],
-            )
-        )
-
-    newobj[
-        padding_width[0] : padding_width[0] + nbz,
-        padding_width[2] : padding_width[2] + nby,
-        padding_width[4] : padding_width[4] + nbx,
-    ] = array
-
-    if debugging:
-        gu.multislices_plot(
-            array=newobj,
-            sum_frames=False,
-            plot_colorbar=True,
-            vmin=0,
-            vmax=1,
-            title="Array after padding",
-        )
-    return newobj
